Move models from news api entry to models dedicated package
